Add public API to read and write HDF5.
* DataSetInfo os now in its own namespace.
* HDF5ImageJ now has `hdf5list` method to list contents of HDF5 file
* `hdf5read` and `hdf5write` methods are now more flexible
* HDF5_Reader_Vibez now uses abstracted DataSetInfo class
