+++ .gitignore
@@ -5,3 +5,7 @@
 /.classpath
 /.project
 /.settings/
+
+# IDEA #
+/.idea/
+*.iml
